Split kedqr command into hash and img
