+++ src/bin/cli/kedqr/mod.rs
@@ -1,12 +1,11 @@
-use catalyst_toolbox::kedqr::{KeyQrCode, QrPin};
-use chain_crypto::bech32::Bech32;
-use chain_crypto::{Ed25519Extended, SecretKey};
-use std::{
-    error::Error,
-    fs::OpenOptions,
-    io::{BufRead, BufReader},
-    path::PathBuf,
-};
+mod hash;
+mod img;
+
+use catalyst_toolbox::kedqr::QrPin;
+pub use hash::generate_hash;
+pub use img::generate_qr;
+use std::error::Error;
+use std::path::PathBuf;
 use structopt::StructOpt;
 
 /// QCode CLI toolkit
@@ -22,50 +21,23 @@ pub struct QrCodeCmd {
     /// Pin code. 4-digit number is used on Catalyst.
     #[structopt(short, long, parse(try_from_str))]
     pin: QrPin,
+
+    #[structopt(flatten)]
+    opts: QrCodeOpts,
 }
 
 impl QrCodeCmd {
     pub fn exec(self) -> Result<(), Box<dyn Error>> {
-        let QrCodeCmd { input, output, pin } = self;
-        // open input key and parse it
-        let key_file = OpenOptions::new()
-            .create(false)
-            .read(true)
-            .write(false)
-            .append(false)
-            .open(&input)
-            .expect("Could not open input file.");
-
-        let mut reader = BufReader::new(key_file);
-        let mut key_str = String::new();
-        let _key_len = reader
-            .read_line(&mut key_str)
-            .expect("Could not read input file.");
-        let sk = key_str.trim_end().to_string();
-
-        let secret_key: SecretKey<Ed25519Extended> =
-            SecretKey::try_from_bech32_str(&sk).expect("Malformed secret key.");
-        // use parsed pin from args
-        let pwd = pin.password;
-        // generate qrcode with key and parsed pin
-        let qr = KeyQrCode::generate(secret_key, &pwd);
-        // process output
-        match output {
-            Some(path) => {
-                // save qr code to file, or print to stdout if it fails
-                let img = qr.to_img();
-                if let Err(e) = img.save(path) {
-                    println!("Error: {}", e);
-                    println!();
-                    println!("{}", qr);
-                }
-            }
-            None => {
-                // prints qr code to stdout when no path is specified
-                println!();
-                println!("{}", qr);
-            }
+        match self.opts {
+            QrCodeOpts::Hash => generate_hash(self.input, self.output, self.pin),
+            QrCodeOpts::Img => generate_qr(self.input, self.output, self.pin),
         }
-        Ok(())
     }
+}
+
+#[derive(Debug, PartialEq, StructOpt)]
+#[structopt(rename_all = "kebab-case")]
+pub enum QrCodeOpts {
+    Img,
+    Hash,
 }
